pyproject.toml
[build-system]
requires = ["setuptools>=70.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sdn-testkit"
version = "0.2.0"
description = "SDN Test Automation Framework — pytest-based testing for software-defined networks"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [{ name = "SDN Test Team" }]
classifiers = [
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Framework :: Pytest",
]
dependencies = [
    "netmiko>=4.3,<5",
    "nornir>=3.4,<4",
    "nornir-netmiko>=1.0,<2",
    "nornir-utils>=0.2,<1",
    "nornir-napalm>=0.4,<1",
    "requests>=2.31,<3",
    "pyyaml>=6.0.2,<7",
    "pydantic>=2.5,<3",
    "jinja2>=3.1,<4",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0,<10",
    "pytest-xdist>=3.5,<4",
    "pytest-html>=4.1,<5",
    "pytest-timeout>=2.3,<3",
    "pytest-rerunfailures>=15,<16",
    "responses>=0.25,<1",
    "ruff>=0.4,<1",
    "mypy>=1.10,<2",
    "pre-commit>=3.7,<4",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "unit: Pure unit tests — no network or controller access (deselect: '-m not unit')",
    "integration: Integration tests — require controller API or mock devices",
    "e2e: End-to-end topology tests — require lab fabric or virtual topology",
    "topology(name): Bind test to a named topology bundle",
    "slow: Tests that take >30s",
    "smoke: Quick sanity checks for CI gating",
]
addopts = [
    "--strict-markers",
    "--tb=short",
    "-v",
    "--timeout=120",
    "--import-mode=importlib",
]
filterwarnings = [
    "ignore::DeprecationWarning:paramiko.*",
]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
